+++ src/app/Model/varietal.ts
@@ -1,0 +1,11 @@
+export class Varietal {
+    VarietalID: number|undefined;
+    Name: string|undefined;
+    Description: string|undefined;
+  
+    constructor() {
+      this.VarietalID = 0;
+      this.Name = '';
+      this.Description = '';
+    }
+  }
